add alchemical-water-cluster notebook
motivation: identify and debug issues using a cheaper test system than a full protein-ligand-binding system...

here, we gradually turn off electrostatic and steric interactions of one water molecule in a cluster of 20 TIP3P water molecules.

an initial test indicated that 0.1ps was too short a protocol (stddev(w_protocol) was approximately 25 kT, so EXP-estimated free energy changes were pretty noisy), so here the protocol is 1ps.

the parameter we are varying is the stepsize dt, which takes values in [0.5,1,1.5,2,2.5,3,3.5,4] femtoseconds. the collision rate is kept constant at 1ps, and the total protocol length (N_steps * dt) is held constant as well. this means that at the smallest dt, we have to take the most computational steps (N_steps = 2000) and at the largest dt, we have to take the fewest steps (N_steps = 250).

seems to encounter NaN's a lot more than anticipated (~10-20% of trajectories encounter NaNs even at 0.5fs, and about ~90% of trajectories encounter NaNs at 4fs, but equilibrium sampling for this system and integrator is stable up to around 8fs)...

additionally, the issue we observed for the CDK2 results (where `mean(w_shadow)` is large at small stepsizes) is visible also here...